% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddText.R
\name{AddText.semantic.twitter}
\alias{AddText.semantic.twitter}
\title{Add columns containing text data to twitter semantic network dataframes}
\usage{
\method{AddText}{semantic.twitter}(net, data, hashtags = FALSE, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{data}{A dataframe generated by \code{Collect}.}

\item{hashtags}{Logical. Add tweet hashtags to dataframes. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges} including columns
containing text data.
}
\description{
Add columns containing text data to twitter semantic network dataframes
}
