% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddUserData.R
\name{AddUserData.actor.twitter}
\alias{AddUserData.actor.twitter}
\title{Supplement twitter actor network by adding user profile attributes to nodes}
\usage{
\method{AddUserData}{actor.twitter}(net, data, rmMisc = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{data}{A dataframe generated by \code{Collect}.}

\item{rmMisc}{Logical. Remove miscellaneous user data columns such as user profile colors and other visual
elements. Default is \code{TRUE}.}

\item{verbose}{Logical. Output additional information. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes}, \code{$edges}. Nodes include
columns for additional user profile data and metrics. Referenced users for which no data was found are returned in an
object attribute named \code{missing_users}.
}
\description{
Network is supplemented with additional downloaded user information applied as actor node attributes.
}
\note{
Using the standard twitter API this function is limited to collecting profiles of 90000 users per 15 mins
before hitting the rate limit. It does not wait and retry upon hitting rate limit.
}
\examples{
\dontrun{
# add user info to a twitter actor network
actor_net <- twitter_data |>
  Create("actor") |>
  AddUserData(twitter_data)
}

}
