% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVideoData.R, R/zzz.R
\name{AddVideoData}
\alias{AddVideoData}
\alias{add_videos}
\title{Add columns of video information to network dataframes}
\usage{
AddVideoData(net, youtubeAuth = NULL, ..., writeToFile = FALSE, verbose = TRUE)

add_videos(net, youtubeAuth = NULL, ..., writeToFile = FALSE, verbose = TRUE)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{youtubeAuth}{YouTube Authenticate object.}

\item{...}{Additional parameters passed to function.}

\item{writeToFile}{Logical. Write data to file. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information. Default is \code{TRUE}.}
}
\value{
Network as a named list of three dataframes containing \code{$nodes}, \code{$edges} and \code{$videos} nodes
and edges include columns for additional video data.
}
\description{
Network is supplemented with additional downloaded video information.
}
\note{
Only supports YouTube actor networks. Refer to \code{\link{AddVideoData.actor.youtube}}.
}
