% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.R, R/Authenticate.reddit.R,
%   R/Authenticate.twitter.R, R/Authenticate.youtube.R
\name{Authenticate}
\alias{Authenticate}
\alias{Authenticate.reddit}
\alias{Authenticate.twitter}
\alias{Authenticate.youtube}
\title{Create a credential to access social media APIs}
\usage{
Authenticate(socialmedia, ...)

\method{Authenticate}{reddit}(socialmedia, ...)

\method{Authenticate}{twitter}(socialmedia, appName, apiKey, apiSecret,
  accessToken, accessTokenSecret, useCachedToken = TRUE, ...)

\method{Authenticate}{youtube}(socialmedia, apiKey, ...)
}
\arguments{
\item{socialmedia}{Character string. Identifier for social media API to authenticate with.\cr
Supports: \code{"twitter"}, \code{"youtube"}, \code{"reddit"}.}

\item{...}{Additional parameters for authentication appropriate to \code{socialmedia} identifier.}

\item{appName}{Character string. Specifies the registered app name associated with API keys.}

\item{apiKey}{Character string. Specifies the app API key used for authentication.}

\item{apiSecret}{Character string. Specifies the app API secret.}

\item{accessToken}{Character string. Specifies the app access token.}

\item{accessTokenSecret}{Character string. Specifies the app access token secret.}

\item{useCachedToken}{Logical. Use cached API token if found otherwise creates one. Default is \code{TRUE}.}
}
\value{
A \code{credential} object with authentication information.
}
\description{
\code{Authenticate} creates a \code{credential} object that enables R to make authenticated calls to social media
APIs. A \code{credential} object is a S3 object with the authentication-related information such as access tokens
and the information on the social media that grant authentication. \code{Authenticate} is the first step of the
\code{Authenticate}, \code{\link{Collect}} and \code{\link{Create}} workflow.
}
\details{
Twitter authentication uses OAuth and requires Application-user authentication API keys as described here: 
\url{https://developer.twitter.com/en/docs/basics/authentication/overview/oauth}

Youtube authentication uses OAuth2 and requires a Google Developer API key as described here:
\url{https://developers.google.com/youtube/v3/docs/}

Reddit does not require authentication in this version of vosonSML.
}
\seealso{
\code{\link{Collect}}, \code{\link{Create}}
}
\keyword{authenticate}
\keyword{credential}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
