% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.reddit.R
\name{Authenticate.reddit}
\alias{Authenticate.reddit}
\title{Reddit API authentication}
\usage{
\method{Authenticate}{reddit}(socialmedia, ...)
}
\arguments{
\item{socialmedia}{Character string. Identifier for social media API to authenticate, set to \code{"reddit"}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
A \code{credential} object containing a \code{$auth = NULL} value and social media type descriptor 
\code{$socialmedia} set to \code{"reddit"}. Object has the class names \code{"credential"} and \code{"reddit"}.
}
\description{
Reddit does not require authentication in this version of vosonSML.
}
\note{
Even though reddit does not require authentication in this version of vosonSML the \code{Authenticate} 
function must still be called to set the \code{socialmedia} identifier. This is used to route to the appropriate 
social media \code{Collect} function.
}
\examples{
\dontrun{
# reddit authentication
redditAuth <- Authenticate("reddit")
}

}
