% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.twitter2.R
\name{Authenticate.twitter2}
\alias{Authenticate.twitter2}
\title{Twitter API v2 authentication}
\usage{
\method{Authenticate}{twitter2}(socialmedia, consumerKey, consumerSecret, bearerToken, ...)
}
\arguments{
\item{socialmedia}{Character string. Identifier for social media API to authenticate, set to \code{"twitter2"}.}

\item{consumerKey}{Character string. API consumer key to authenticate.}

\item{consumerSecret}{Character string. API consumer secret to authenticate.}

\item{bearerToken}{Character string. API access token to authenticate.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
A \code{credential} object containing a bearer token \code{$bearer} and social media type descriptor
\code{$socialmedia} set to \code{"twitter2"}. Object has the class names \code{"credential"} and \code{"twitter2"}.
}
\description{
Twitter API v2 authentication uses OAuth2 and requires app consumer keys or bearer token.
}
\examples{
\dontrun{
# twitter API v2 consumer keys
twitter2Auth <- Authenticate("twitter2",
  consumerKey = "xxxxxxxx",
  consumerSecret = "xxxxxxxx")

# twitter API v2 bearer token
twitter2Auth <- Authenticate("twitter2",
  bearerToken = "xxxxxxxx")
)
}

}
