% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.twitter.R
\name{Authenticate.twitter_convo}
\alias{Authenticate.twitter_convo}
\title{Twitter API v2 authentication}
\usage{
\method{Authenticate}{twitter_convo}(socialmedia, apiKey, apiSecret, bearer = NULL, ...)
}
\arguments{
\item{socialmedia}{Character string. Identifier for social media API to authenticate, set to \code{"twitter_convo"}.}

\item{apiKey}{Character string. API consumer key to authenticate.}

\item{apiSecret}{Character string. API consumer secret to authenticate.}

\item{bearer}{Character string. Sets bearer token to string value. If NULL retrieves application bearer token from
API. Default is \code{NULL}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\description{
Twitter API v2 conversation authentication currently only supports app bearer tokens.
}
\examples{
\dontrun{
# twitter app keys
twitterAuth <- Authenticate("twitter_convo",
                             apiKey = "xxxxxxxxxxxx",
                             apiSecret = "xxxxxxxxxxxx",
                             bearer = NULL)

# if already have the app bearer token
twitterAuth <- Authenticate("twitter_convo", bearer = "xxxxxxxxxxxx")
}

}
