% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_mastodon.R
\name{authenticate_mastodon}
\alias{authenticate_mastodon}
\title{Mastodon authentication}
\usage{
authenticate_mastodon(
  instance = NULL,
  type = "public",
  browser = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{instance}{Character vector. Mastodon instance to authenticate with. Default is \code{NULL} for no authentication.}

\item{type}{Character vector. Type of authentication, can be \code{"public"} or \code{"user"} Default is \code{"public"}}

\item{browser}{Logical. Launch the default web browser to complete authentication. Default is \code{TRUE}.}

\item{verbose}{Logical. Display verbose messages. Default is \code{FALSE}.}
}
\value{
A named list with an auth token if authenticated.
}
\description{
Mastodon API OAuth authentication.
}
