% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateWithRedditAPI.R
\name{AuthenticateWithRedditAPI}
\alias{AuthenticateWithRedditAPI}
\title{Reddit API Authentication.}
\usage{
AuthenticateWithRedditAPI(app_name, app_key, app_secret, use_token_cache)
}
\arguments{
\item{app_name}{Character string containing the reddit app name associated with the API key.}

\item{app_key}{Character string containing the app key.}

\item{app_secret}{Character string containing the app secret.}

\item{use_token_cache}{Boolean. Use cached authentication token if found.}
}
\value{
Reddit authentication token.
}
\description{
Oauth2 based authentication with the Reddit API that returns an authentication token.
}
\details{
The httr package has a known oauth2 issue with its parameter "use_basic_auth", The default value is set to FALSE
and is missing parameter pass through meaning it can not be set to TRUE as required by reddit oauth2 authentication.
The point patch devtools::install_github("r-lib/httr#485") fixes this issue.
Further information: https://github.com/r-lib/httr/issues/482

Reddit oauth tokens are only valid for one hour and using cached token will subsequently produce 401 errors.
}
