% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateWithTwitterAPI.R
\name{AuthenticateWithTwitterAPI}
\alias{AuthenticateWithTwitterAPI}
\title{Note: this function is DEPRECATED. Please use the \code{\link{Authenticate}} function.}
\usage{
AuthenticateWithTwitterAPI(appName, apiKey, apiSecret, accessToken,
  accessTokenSecret, useCachedToken)
}
\arguments{
\item{appName}{Character string. Specifies the twitter registered app name associated with API keys.}

\item{apiKey}{Character string. Specifies the app 'API key' used for authentication.}

\item{apiSecret}{Character string. Specifies the app 'API secret'.}

\item{accessToken}{Character string. Specifies the app 'access token'.}

\item{accessTokenSecret}{Character string. Specifies the app 'access token secret'.}

\item{useCachedToken}{Logical. If \code{TRUE} uses cached API token if found otherwise creates one.}
}
\value{
twitter_oauth. Returns a twitter oauth token object.
}
\description{
Twitter API authentication
}
\details{
Oauth based authentication using the Twitter API.

In order to collect data from Twitter, the user must first authenticate with Twitter's API. This requires setting up 
an app on Twitter. A useful guide to creating an app can be found in the rtweet documentation: 
https://rtweet.info/articles/auth.html#creating-a-twitter-app
}
\seealso{
\code{\link{Authenticate}}
}
\keyword{authenticate}
\keyword{twitter}
