% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateWithYoutubeAPI.R
\name{AuthenticateWithYoutubeAPI}
\alias{AuthenticateWithYoutubeAPI}
\title{YouTube API Authentication}
\usage{
AuthenticateWithYoutubeAPI(apiKeyYoutube)
}
\arguments{
\item{apiKeyYoutube}{character string specifying your Google Developer API key.}
}
\value{
This is called for its side effect.
}
\description{
OAuth based authentication with the Google API.
}
\details{
In order to collect data from YouTube, the user must first authenticate with Google's Application Programming 
Interface (API). Users can obtain a Google Developer API key at: https://console.developers.google.com.
}
\note{
In the future this function will enable users to save the API key in working directory, and the function will 
automatically look for a locally stored key whenever it is called without \code{apiKeyYoutube} argument.
}
\seealso{
\code{AuthenticateWithFacebookAPI} and \code{AuthenticateWithTwitterAPI} for other ways to collect social 
media data.
}
\author{
Timothy Graham <timothy.graham@anu.edu.au> & Robert Ackland <robert.ackland@anu.edu.au>
}
\keyword{SNA}
\keyword{media}
\keyword{social}
\keyword{youtube}
