% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.R
\name{Collect}
\alias{Collect}
\title{Collect data from social media for generating networks}
\usage{
Collect(credential, ego = FALSE, ...)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate}.}

\item{ego}{Logical. If \code{TRUE} collect ego network data. Currently only supports Instagram.}

\item{...}{Additional parameters for data collection by appropriate to credential \code{socialmedia} type.
Refer to CollectDataFrom* and CollectEgo* functions for more details.
\describe{
  \item{twitter:}{\code{authToken, searchTerm, [searchType, numTweets, includeRetweets, retryOnRateLimit,}\cr
                  \code{writeToFile, verbose, ...]}}
  \item{youtube:}{\code{videoIDs, apiKeyYoutube, [verbose, writeToFile, maxComments]}}
  \item{reddit:}{\code{threadUrls, [waitTime, writeToFile]}}
  \item{instagram:}{\code{tag, n, lat, lng, [distance, folder, mindate, maxdate, verbose, sleep,}\cr
                    \code{writeToFile, waitForRateLimit, credential]}}
  \item{instagram with \code{ego = TRUE}:}{\code{username, userid, [verbose, degreeEgoNet,}\cr
                                           \code{waitForRateLimit, getFollows, credential]}}
  \item{facebook:}{\code{pageName, [rangeFrom, rangeTo, verbose, n, writeToFile, dynamic]}}
}}
}
\value{
A data.frame object of class \code{dataSource.*} that can be used with \code{Create}.
}
\description{
This function collects data from social media APIs, and structures the data into a data frame of class
\code{dataSource.*}, ready for creating networks for further analysis. \code{Collect} is the second step of the
\code{Authenticate}, \code{Collect}, \code{Create} workflow. This function is a convenient UI wrapper to the core
CollectDataFrom* family of functions.
}
\examples{
\dontrun{
require(magrittr)

## youtube actor network example

myYoutubeAPIKey <- "xxxxxxxxxxxxxxxxxxxxxx"
listYoutubeVideoIDs <- c("W2GZFeYGU3s", "mL27TAJGlWc")

myActorNetwork <- Authenticate("youtube", apiKey = myYoutubeAPIKey) \%>\%
  Collect(videoIDs = listYoutubeVideoIDs) \%>\% Create("actor")

## instagram ego network example

myInstaAppID <- "xxxxxxxxxxx"
myInstaAppSecret <- "xxxxxxxxxxxxxxxxxxxxxx"
listInstaUsernames <- c("senjohnmccain", "obama")

myEgoNetwork <- Authenticate("instagram", appID = myInstaAppID, appSecret = myInstaAppSecret) \%>\%
  Collect(ego = TRUE, username = listInstaUsernames) \%>\% Create("ego")

## facebook bimodal network example

myFacebookAppID <- "xxxxxxxxxxx"
myFacebookAppSecret <- "xxxxxxxxxxxxxxxxxxxxxx"

myBimodalNetwork <- Authenticate("Facebook", appID = myFacebookAppID,
                                 appSecret = myFacebookAppSecret) \%>\%
  SaveCredential("FBCredential.RDS") \%>\%
  Collect(pageName = "StarWars", rangeFrom = "2015-03-01", rangeTo = "2015-03-02",
          writeToFile = FALSE) \%>\%
  Create("bimodal")

## facebook dynamic network example

myDynamicNetwork <- LoadCredential("FBCredential.RDS") \%>\%
  Collect(pageName = "StarWars", rangeFrom = "2015-03-01", rangeTo = "2015-03-02",
          writeToFile = FALSE) \%>\%
  Create("dynamic")
}

}
\seealso{
\code{Authenticate}, \code{Create}
}
\keyword{collect}
\keyword{facebook}
\keyword{instagram}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
