% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.timeline.twitter.R
\name{Collect.timeline.twitter}
\alias{Collect.timeline.twitter}
\title{Collect tweet data from twitter timelines}
\usage{
\method{Collect}{timeline.twitter}(
  credential,
  endpoint,
  users = c(),
  numTweets = 100,
  writeToFile = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate} with class name \code{"twitter"}.}

\item{endpoint}{API endpoint.}

\item{users}{Character vector. Specifies one or more twitter users. Can be user names, user ids or a mixture.}

\item{numTweets}{Numeric. Specifies how many tweets to be collected per user. Defaults is \code{100}.}

\item{writeToFile}{Logical. Write collected data to file. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the data collection. Default is \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[rtweet:get_timeline]{rtweet::get_timeline}}
  \describe{
    \item{\code{parse}}{Logical, indicating whether to return parsed
(data.frames) or nested list object. By default, \code{parse =
TRUE} saves users from the time [and frustrations] associated
with disentangling the Twitter API return objects.}
    \item{\code{check}}{Logical indicating whether to remove check available
rate limit. Ensures the request does not exceed the maximum
remaining number of calls.  Defaults to TRUE.}
  }}
}
\value{
A data.frame object with class names \code{"datasource"} and \code{"twitter"}.
}
\description{
This function collects user timeline tweets and structures the data into a dataframe with the class
names \code{"datasource"} and \code{"twitter"}. A maximum of 3,200 tweets can be retrieved per user.
}
