% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.twitter_convo.R
\name{Collect.twitter_convo}
\alias{Collect.twitter_convo}
\title{Collect tweet data from twitter conversation}
\usage{
\method{Collect}{twitter_convo}(credential, tweet_id, writeToFile = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate} with class name \code{"twitter"}.}

\item{tweet_id}{Numeric. Tweet id of any tweet in the twitter conversation to collect.}

\item{writeToFile}{Logical. Write collected data to file. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the data collection. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
A data.frame object with class names \code{"datasource"} and \code{"twitter_convo"}.
}
\description{
This function collects conversation tweets.
}
\note{
bearer rate-limits (per app):
  300 tweet lookup requests (300*100 = 30,000 tweets) per 15 min window
  450 recent search requests (450*100 = 45,000 tweets) per 15 min window
}
