% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.youtube.R
\name{Collect.youtube}
\alias{Collect.youtube}
\title{Collect youtube comments data for generating different types of networks}
\usage{
\method{Collect}{youtube}(credential, videoIDs, verbose = FALSE,
  writeToFile = FALSE, maxComments = 1e+13, ...)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate} with class name \code{"youtube"}.}

\item{videoIDs}{Character vector. Specifies one or more youtube video IDs. For example, if the video URL is 
\code{https://www.youtube.com/watch?v=xxxxxxxxxxx} then use \code{videoIDs = c("xxxxxxxxxxx")}. For multiple videos 
the function \code{\link{GetYoutubeVideoIDs}} can be used to create a vector object suitable as input for videoIDs.}

\item{verbose}{Logical. Output additional information about the data collection. Default is \code{FALSE}.}

\item{writeToFile}{Logical. Write collected data to file. Default is \code{FALSE}.}

\item{maxComments}{Numeric integer. Specifies how many top-level comments to collect from each video. This value 
does not consider replies to top-level comments. The total number of comments returned for a video will usually be 
greater than \code{maxComments} depending on the number of reply comments present.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
A data.frame object with class names \code{"datasource"} and \code{"youtube"}.
}
\description{
This function collects public comments data for one or more youtube videos using the Google Data API. 
It then structures the data into a dataframe of class \code{"datasource"} and \code{"youtube"} for use in subsequent 
network creation with the \code{\link{Create}} function.
}
\note{
Due to specifications of the Google Data API it is currently not efficient to specify the exact number of 
comments to return from the API using \code{maxComments} parameter. The \code{maxComments} parameter is applied to
top-level comments only and not the replies to these comments. As such the number of comments collected is usually 
greater than expected. For example, if \code{maxComments} is set to 10 and one of the videos 10 top-level comments 
has 5 reply comments then the total number of comments collected will be 15 for that video. Comments data for 
multiple youtube videos can be requested in a single operation, \code{maxComments} is applied to each individual 
video and not the combined total of comments.
}
\examples{
\dontrun{
# create a list of youtube video ids to collect on
videoIDs <- GetYoutubeVideoIDs(c("https://www.youtube.com/watch?v=xxxxxxxx", 
                                 "https://youtu.be/xxxxxxxx"))

# collect approximately 200 threads/comments for the youtube video
youtubeData <- youtubeAuth \%>\% 
  Collect(videoIDs = videoIDs, writeToFile = TRUE, verbose = FALSE, maxComments = 200)
}

}
