% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_mastodon_timelines.R
\name{collect_mastodon_timeline}
\alias{collect_mastodon_timeline}
\title{Mastodon timeline collection}
\usage{
collect_mastodon_timeline(
  auth = NULL,
  instance = NULL,
  account = NULL,
  hashtag = NULL,
  local = FALSE,
  limit = 100,
  anonymous = TRUE,
  retryonratelimit = TRUE,
  extract = TRUE,
  file = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{auth}{List. Mastodon authentication. If is \code{NULL} anonymous API access will be assumed.}

\item{instance}{Character string. Server to collect timeline posts from. If left \code{NULL} auth token instance will
be used.}

\item{account}{Character string. Collect posts from a specified account username or id. Default is \code{NULL}.}

\item{hashtag}{Character string. Specifies a hashtag filter e.g \code{"rstats"}. Default is \code{NULL} to return all
unfiltered timeline posts.}

\item{local}{Logical. Search only the local server timeline or global federated timeline. Default is \code{FALSE}.}

\item{limit}{Numeric. Specify the collection limit for timeline posts. Default is \code{100}.}

\item{anonymous}{Logical. Collect public posts without authenticating. Default is \code{TRUE}.}

\item{retryonratelimit}{Logical. When the API rate-limit is reached should the collection wait and resume when it
resets. Default is \code{TRUE}.}

\item{extract}{Logical. Extract metadata into seperate dataframes. Default is \code{TRUE}.}

\item{file}{Logical or Character. Write the collected data to file. If \code{TRUE} will generate a file name and
write to default path. A file path can be specified. Default is \code{FALSE}.}

\item{verbose}{Logical. Display verbose messages. Default is \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[rtoot:get_timeline_hashtag]{rtoot::get_timeline_hashtag}}
  \describe{
    \item{\code{only_media}}{logical, Show only statuses with media attached?}
    \item{\code{min_id}}{character, Return results immediately newer than this id}
    \item{\code{parse}}{logical, if \code{TRUE}, the default, returns a tibble. Use \code{FALSE}  to return the "raw" list corresponding to the JSON returned from the Mastodon API.}
  }}
}
\value{
A named list of tibbles containing mastodon collection data
}
\description{
Mastodon collection of instance timeline data.
}
