% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataReddit.R
\name{CollectDataReddit}
\alias{CollectDataReddit}
\title{Collect reddit data.}
\usage{
CollectDataReddit(oauth2_token, thread_urls, ua, write_to_file)
}
\arguments{
\item{oauth2_token}{A Token object returned from reddit oauth2 authentication.}

\item{thread_urls}{A vector of reddit thread url's to collect data from.}

\item{ua}{Character string containing a user-agent string to use in the get request.}

\item{write_to_file}{Boolean. If the data should be written to file.}
}
\value{
A data frame object of class \code{dataSource.reddit} that can be used for creating unimodal 
networks (\code{CreateActorNetwork}).
}
\description{
Uses reddit thread url's to collect user and comment data.
}
