% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataTwitter.R
\name{CollectDataTwitter}
\alias{CollectDataTwitter}
\title{Note: this function is DEPRECATED. Please use the \code{\link{Collect}} function.}
\usage{
CollectDataTwitter(authToken, searchTerm, searchType, numTweets,
  includeRetweets, retryOnRateLimit, writeToFile, verbose, ...)
}
\arguments{
\item{authToken}{Twitter oauth token created by rtweet.}

\item{searchTerm}{Character string. Specifies a search term or phrase (e.g. "Australian politics") or hashtag (e.g. 
"#auspol"). Many query operators are available - see the Twitter documentation for more information: 
https://dev.twitter.com/rest/public/search}

\item{searchType}{Character string. Returns filtered tweets as per search type \code{recent}, \code{mixed} or 
\code{popular}. Default type is \code{recent}.}

\item{numTweets}{Numeric. Specifies how many tweets to be collected. Defaults is \code{100}.}

\item{includeRetweets}{Logical. Specifies if the search should filter out retweets. Defaults is \code{TRUE}.}

\item{retryOnRateLimit}{Logical. Default is \code{FALSE}.}

\item{writeToFile}{Logical. If \code{TRUE} then the data is saved to file in current working directory (RDS format), 
with filename denoting current system time and \code{searchTerm}. Default is \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} then this function will output runtime information to the console as it 
computes. Useful diagnostic tool for long computations. Default is \code{FALSE}.}

\item{...}{Additional parameters to pass to the rtweet \code{search_tweets} function.}
}
\value{
A data frame object of class \code{dataSource.twitter} that can be used for creating unimodal networks 
(\code{CreateActorNetwork}), bimodal networks (\code{CreateBimodalNetwork}), and semantic networks 
(\code{CreateSemanticNetwork}).
}
\description{
Collect data from Twitter for generating different types of networks
}
\details{
This function collects data from Twitter based on hashtags or search terms, and structures the data into a data 
frame of class \code{dataSource.twitter}, ready for creating networks for further analysis.

\code{CollectDataTwitter} collects public 'tweets' from Twitter using the Twitter API.

The function then finds and maps the relationships of entities of interest in the data (e.g. users, terms, hashtags)
, and structures these relationships into a data frame format suitable for creating unimodal networks 
(\code{CreateActorNetwork}), bimodal networks (\code{CreateBimodalNetwork}), and semantic networks 
(\code{CreateSemanticNetwork}).

The maximum number of tweets for a single call of \code{CollectDataTwitter} is 1500.

Language support is available, using the \code{language} parameter. The user can restrict tweets returned to a 
particular language, using the ISO 639-1 code. For example, restricting to English would use \code{language="en"}. 
The full list of codes is available here: https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes.

A variety of query operators are available through the Twitter API. For example, "love OR hate" returns any tweets 
containing either term (or both). For more information see the Twitter API documentation (under the heading
'Query Operators'): https://dev.twitter.com/rest/public/search
}
\note{
Supported network types: \code{actor}, \code{bimodal}, \code{semantic}

Data generated using this function is *not* suitable for dynamic networks.
}
\seealso{
\code{Collect}
}
\keyword{collect}
\keyword{twitter}
