% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.actor.mastodon.R
\name{Create.actor.mastodon}
\alias{Create.actor.mastodon}
\title{Create mastodon actor network}
\usage{
\method{Create}{actor.mastodon}(datasource, type, inclMentions = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"mastodon"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"actor"}.}

\item{inclMentions}{Logical. Create edges for users mentioned or tagged in posts. Default is \code{TRUE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a mastodon actor network from posts. Mastodon users who have posted are actor nodes. The
created network is directed with edges representing replies.
}
\examples{
\dontrun{
# create a mastodon actor network
actor_net <- mastodon_data |> Create("actor")
}

}
