% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.actor.R, R/Create.actor.reddit.R,
%   R/Create.actor.twitter.R, R/Create.actor.youtube.R
\name{Create.actor}
\alias{Create.actor}
\alias{Create.actor.reddit}
\alias{Create.actor.twitter}
\alias{Create.actor.youtube}
\title{Create actor network from social media data}
\usage{
\method{Create}{actor}(datasource, type, ...)

\method{Create}{actor.reddit}(datasource, type, weightEdges = FALSE,
  textData = FALSE, cleanText = TRUE, writeToFile = FALSE, ...)

\method{Create}{actor.twitter}(datasource, type, writeToFile = FALSE,
  verbose = FALSE, ...)

\method{Create}{actor.youtube}(datasource, type, writeToFile = FALSE,
  ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{datasource} and \code{socialmedia} class attributes.}

\item{type}{Character string. Type of network to be created.}

\item{...}{Additional parameters to pass to the network creation method.}

\item{weightEdges}{Logical. Combine and weight directed network edges. Default is \code{FALSE}.}

\item{textData}{Logical. Include comment text as an edge attributes of returned igraph network. 
Cannot be used with the \code{weightEdges} parameter. Default is \code{FALSE}.}

\item{cleanText}{Logical. Remove non-alphanumeric, non-punctuation, and non-space characters from the comment text 
data used as edge attributes. Default is \code{TRUE}.}

\item{writeToFile}{Logical. Save network data to a file in the current working directory. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}
}
\description{
Create actor network from social media data

Create reddit actor network

Create twitter actor network

Create youtube actor network
}
\note{
Supported data sources: \code{twitter}, \code{youtube}, \code{reddit}
}
\seealso{
\code{\link{Create}}
}
\keyword{actor}
\keyword{create}
\keyword{network}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
