% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.bimodal.R, R/Create.bimodal.twitter.R
\name{Create.bimodal}
\alias{Create.bimodal}
\alias{Create.bimodal.twitter}
\title{Create bimodal networks from social media data}
\usage{
\method{Create}{bimodal}(datasource, type, ...)

\method{Create}{bimodal.twitter}(datasource, type,
  removeTermsOrHashtags = NULL, writeToFile = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{datasource}{A dataframe of class \code{datasource} and class of a social media type.}

\item{type}{Character string. Type of network to be created.}

\item{...}{Additional parameters to pass to the network creation method.}

\item{removeTermsOrHashtags}{Character string. Default is none. Otherwise this argument specifies which terms or 
hashtags (i.e. vertices with matching 'name') should be removed from the bimodal network. This is useful to remove
the search term or hashtag that was used to collect the data (i.e. remove the corresponding vertex in the graph). 
For example, a value of "#auspol" means that if there is a vertex with the exact name "#auspol" then this vertex 
will be removed.}

\item{writeToFile}{Logical. If \code{TRUE} then the network is saved to file in current working directory (graphml 
format), with filename denoting the current datetime and the type of network.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}
}
\value{
A twitter bimodal network as igraph object.
}
\description{
This function creates a directed and weighted bimodal network from a data frame of class \code{dataSource} (as 
created by the CollectData functions).
}
\note{
Supported data sources: \code{twitter}
}
\seealso{
\code{\link{Create}}
}
\keyword{bimodal}
\keyword{create}
\keyword{network}
\keyword{twitter}
