% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.bimodal.twitter.R
\name{Create.bimodal.twitter}
\alias{Create.bimodal.twitter}
\title{Create twitter bimodal network}
\usage{
\method{Create}{bimodal.twitter}(datasource, type,
  removeTermsOrHashtags = NULL, writeToFile = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{datasource} and \code{socialmedia} class attributes.}

\item{type}{Character string. Type of network to be created.}

\item{removeTermsOrHashtags}{Character string. Default is none. Otherwise this argument specifies which terms or 
hashtags (i.e. vertices with matching 'name') should be removed from the bimodal network. This is useful to remove
the search term or hashtag that was used to collect the data (i.e. remove the corresponding vertex in the graph). 
For example, a value of "#auspol" means that if there is a vertex with the exact name "#auspol" then this vertex 
will be removed.}

\item{writeToFile}{Logical. If \code{TRUE} then the network is saved to file in current working directory (graphml 
format), with filename denoting the current datetime and the type of network.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}

\item{...}{Additional parameters to pass to the network creation method.}
}
\value{
A twitter bimodal network as igraph object.
}
\description{
Creates a bimodal network from collected tweets.
}
