% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.R
\name{Create}
\alias{Create}
\title{Create networks from social media data}
\usage{
Create(datasource, type, ...)
}
\arguments{
\item{datasource}{Collected social media data of class \code{"datasource"} and \code{socialmedia}.}

\item{type}{Character string. Type of network to be created, can be \code{actor}, \code{bimodal} or \code{semantic}.}

\item{...}{Optional parameters to pass to functions providied by supporting R packages that are used for social media 
API network creation.}
}
\description{
This function creates networks from social media data as produced from \code{Collect}. \code{Create} is 
the final step of the \code{Authenticate}, \code{Collect}, \code{Create} workflow.

Refer to \code{\link{Create.actor.twitter}}, \code{\link{Create.actor.youtube}} and 
\code{\link{Create.actor.reddit}} for parameters and usage.
}
