% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mastodon_network.R
\name{create_mastodon_network}
\alias{create_mastodon_network}
\title{Mastodon network creation}
\usage{
create_mastodon_network(data = NULL, type = "activity", verbose = FALSE, ...)
}
\arguments{
\item{data}{List. Named list containing data.}

\item{type}{Character. Type of network to create.}

\item{verbose}{Logical. Display verbose messages. Default is \code{FALSE}.}

\item{...}{Additional parameters to pass to graph function.}
}
\value{
A named list with network as a tidygraph network graph object
}
\description{
Mastodon networks from collected data.
}
