% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.activity.twitter.R
\name{create_network.activity.twitter}
\alias{create_network.activity.twitter}
\title{Create twitter activity network}
\usage{
\method{create_network}{activity.twitter}(datasource, type, rm_node_types = NULL, verbose = TRUE, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"twitter"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"activity"}.}

\item{rm_node_types}{Character vector. List of tweet types to remove from network. Options are \code{"tweet"},
\code{"retweet"}, \code{"reply"} and \code{"quote"}. Default is \code{NULL}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{TRUE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a twitter activity network from collected tweets. Nodes are tweets and directed edges represent
the relationship of tweets to one another. For example, there is a directed edge from a quote tweet towards the
tweet that was quoted. Stand-alone tweets that are not replies, retweets or quote tweets have no relation to others
and will be isolates.
}
\examples{
\dontrun{
# create a twitter activity network with retweets removed
activity_net <- twitter_data \%>\%
  create_network("activity", rm_node_types = c("retweet"))
  
# network nodes and edges
# activity_net$nodes
# activity_net$edges
}

}
