% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.activity.web.R
\name{create_network.activity.web}
\alias{create_network.activity.web}
\title{Create web activity network}
\usage{
\method{create_network}{activity.web}(datasource, type, verbose = TRUE, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"web"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"activity"}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{TRUE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a web page activity network from pages. Nodes are web pages.
}
\examples{
\dontrun{
# create a web activity network graph
activity_net <- web_data \%>\%
  create_network("activity")

# network nodes and edges
# activity_net$nodes
# activity_net$edges
}

}
