% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.actor.reddit.R
\name{create_network.actor.reddit}
\alias{create_network.actor.reddit}
\title{Create reddit actor network}
\usage{
\method{create_network}{actor.reddit}(datasource, type, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"reddit"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"actor"}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a reddit actor network from thread comments on subreddits. Users who have commented on a thread 
are actor nodes and comment replies to each other are represented as directed edges.
}
\examples{
\dontrun{
# create a reddit actor network
actor_net <- reddit_data \%>\%
  create_network("actor")

# network nodes and edges
# actor_net$nodes
# actor_net$edges
}

}
