% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.actor.twitter.R
\name{create_network.actor.twitter}
\alias{create_network.actor.twitter}
\title{Create twitter actor network}
\usage{
\method{create_network}{actor.twitter}(
  datasource,
  type,
  rm_edge_types = NULL,
  incl_mentions = TRUE,
  incl_rt_mentions = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"twitter"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"actor"}.}

\item{rm_edge_types}{Character vector. List of edge types to remove from network. Options are \code{"tweet"},
\code{"retweet"}, \code{"reply"} and \code{"quote"}. Default is \code{NULL}.}

\item{incl_mentions}{Logical. Create edges for users mentioned or tagged in tweets. Default is \code{TRUE}.}

\item{incl_rt_mentions}{Logical. Create edges for users mentioned or tagged in retweets. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{TRUE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a twitter actor network from tweets returned from the twitter search query. Twitter users who 
have tweeted / retweeted or been mentioned in a tweet are actor nodes. The created network is directed with edges 
of different types representing retweets, quote tweets, mentions and replies to other users. Users who have tweeted 
without relations to other users will appear in the network graph as isolate nodes.
}
\note{
When creating twitter actor networks, a network with additional user information can be generated using the
\code{\link{AddUserData}} function. Additional calls can be made to the twitter API to get information
about users that were identified as nodes during network creation but did not tweet (meaning no user profile 
information was initially collected for them).
}
\examples{
\dontrun{
# create a twitter actor network excluding retweet, quote tweets and mention edges
actor_net <- twitter_data \%>\%
  create_network("actor", rm_edge_types = c("retweet", "quote"), incl_mentions = FALSE)

# network nodes and edges
# actor_net$nodes
# actor_net$edges
}

}
