% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.actor.youtube.R
\name{create_network.actor.youtube}
\alias{create_network.actor.youtube}
\title{Create youtube actor network}
\usage{
\method{create_network}{actor.youtube}(datasource, type, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"youtube"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"actor"}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a youtube actor network from comment threads on youtube videos. Users who have made comments to 
a video (top-level comments) and users who have replied to those comments are actor nodes. The comments are 
represented as directed edges between the actors. The video id is also included as an actor node, representative of 
the videos publisher with top-level comments as directed edges towards them.
}
\examples{
\dontrun{
# create a youtube actor network
actor_net <- youtube_data \%>\% create_network("actor")

# network nodes and edges
# actor_net$nodes
# actor_net$edges
}

}
