% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{create_network}
\alias{create_network}
\title{Create networks from social media data}
\usage{
create_network(datasource, type, ...)
}
\arguments{
\item{datasource}{Collected social media data of class \code{"datasource"} and \code{socialmedia}.}

\item{type}{Character string. Type of network to be created, can be \code{"activity"}, \code{"actor"},
\code{"twomode"} or \code{"semantic"}.}

\item{...}{Optional parameters to pass to functions provided by supporting R packages that are used for social 
media network creation.}
}
\description{
This function creates networks from social media data as produced from \code{\link{Collect}}. 
\code{create_network} is the final step of the \code{\link{Authenticate}}, \code{\link{Collect}} and 
\code{create_network} workflow.

There are four types of networks that can be created from collected data: \code{activity}, \code{actor}, 
\code{twomode} or \code{semantic}.

For \code{activity} networks refer to \code{\link{create_network.activity.twitter}}, 
\code{\link{create_network.activity.youtube}} 
and \code{\link{create_network.activity.reddit}} for parameters and usage.

For \code{actor} networks refer to \code{\link{create_network.actor.twitter}}, 
\code{\link{create_network.actor.youtube}} and \code{\link{create_network.actor.reddit}}.

For \code{twomode} and \code{semantic} networks refer to \code{\link{create_network.twomode.twitter}} and 
\code{\link{create_network.semantic.twitter}} functions for parameters and usage respectively.
}
