% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.twomode.twitter.R
\name{create_network.twomode.twitter}
\alias{create_network.twomode.twitter}
\title{Create twitter 2-mode network}
\usage{
\method{create_network}{twomode.twitter}(
  datasource,
  type,
  rm_nodes = NULL,
  weighted = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"twitter"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"twomode"}.}

\item{rm_nodes}{Character vector. List of \code{@Users} and \code{#hashtags} to remove from the 2-mode network.
For example, this parameter could be used to remove the user or hashtag that was used to collect the data by
removing any nodes with matching name. Default is \code{NULL}.}

\item{weighted}{Logical. Add weights to network edges. If set to \code{FALSE} tweet \code{tweet_id} and 
\code{created_at} fields will be preserved for edges in the dataframe. Default is \code{TRUE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a 2-mode network from tweets returned from the twitter search query. In this network there are 
two types of nodes, twitter users who authored or were mentioned in collected tweets and hashtags found within
tweets. Network edges represent a users tweets that contain hashtags or mention users screen names.

The creation of twitter 2-mode networks requires text processing and the tokenization of tweets. As such
this function requires the additional installation of the \pkg{tidytext} package to achieve this.
}
\examples{
\dontrun{
# twitter 2-mode network creation additionally requires the tidytext package
# for working with text data
install.packages("tidytext")

# create a twitter 2-mode network graph removing the hashtag '#auspol'
2mode_net <- twitter_data \%>\% 
  create_network("twomode", rm_nodes = c("#auspol"), verbose = TRUE)

# network nodes and edges
# 2mode_net$nodes
# 2mode_net$edges
}

}
