% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_graph.R
\name{create_network_graph}
\alias{create_network_graph}
\title{Create a network graph object}
\usage{
create_network_graph(
  network,
  type = "tidygraph",
  file = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{network}{Named list. Network object returned from create network function.}

\item{type}{Character. Type of graph object to return, can be \code{"tidygraph"} or \code{"igraph"}. Default is
\code{"tidygraph"}.}

\item{file}{Logical or Character. Write the collected data to file. If \code{TRUE} will generate a file name and
write to default path. A file path can be specified. Default is \code{FALSE}.}

\item{verbose}{Logical. Display verbose messages. Default is \code{FALSE}.}

\item{...}{Additional parameters to pass to \pkg{igraph} or \pkg{tidygraph} graph function.}
}
\value{
A network graph object
}
\description{
Create a specified type of graph object from a tibble of network edges and a tibble of network nodes.
}
