% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateActorNetwork.R,
%   R/CreateActorNetwork.reddit.R, R/CreateActorNetwork.twitter.R,
%   R/CreateActorNetwork.youtube.R
\name{CreateActorNetwork}
\alias{CreateActorNetwork}
\alias{CreateActorNetwork.reddit}
\alias{CreateActorNetwork.twitter}
\alias{CreateActorNetwork.youtube}
\title{Create actor network from social media data}
\usage{
CreateActorNetwork(datasource, ...)

\method{CreateActorNetwork}{reddit}(datasource, weightEdges = FALSE,
  textData = FALSE, cleanText = TRUE, writeToFile = FALSE, ...)

\method{CreateActorNetwork}{twitter}(datasource, writeToFile = FALSE,
  verbose = FALSE, ...)

\method{CreateActorNetwork}{youtube}(datasource, writeToFile = FALSE,
  ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{datasource} and \code{socialmedia} class attributes.}

\item{...}{Additional parameters to pass to the network creation method.}

\item{weightEdges}{Logical. Combine and weight directed network edges. Default is \code{FALSE}.}

\item{textData}{Logical. Include comment text as an edge attributes of returned igraph network. 
Cannot be used with the \code{weightEdges} parameter. Default is \code{FALSE}.}

\item{cleanText}{Logical. Remove non-alphanumeric, non-punctuation, and non-space characters from the comment text 
data used as edge attributes. Default is \code{TRUE}.}

\item{writeToFile}{Logical. Save network data to a file in the current working directory. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}
}
\description{
Create actor network from social media data

Create reddit actor network

Create twitter actor network

Create youtube actor network
}
\note{
Supported \code{datasource} \code{socialmedia}: \code{twitter}, \code{youtube}, \code{reddit}
}
\seealso{
\code{\link{Create}}
}
\keyword{actor}
\keyword{create}
\keyword{network}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
