% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateActorNetwork.reddit.R
\name{CreateActorNetwork.reddit}
\alias{CreateActorNetwork.reddit}
\title{Creates a reddit actor network from collected threads}
\usage{
\method{CreateActorNetwork}{reddit}(x, includeTextData, cleanText,
  writeToFile)
}
\arguments{
\item{x}{a dataframe as vosonSML class object containing collected social network data}

\item{includeTextData}{logical. If the igraph network edges should include the text attribute.}

\item{cleanText}{logical. If non-alphanumeric, non-punctuation, and non-space characters should be removed from the 
included text attribute data.}

\item{writeToFile}{logical. If the igraph network graph should be written to file.}
}
\value{
an igraph object of the actor network
}
\description{
Uses RedditExtractoR::user_network to create an igraph actor network with posts as edge attribute.
}
