% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnhanceNetwork.twitter.R
\name{EnhanceNetwork.twitter}
\alias{EnhanceNetwork.twitter}
\title{Enhances twitter network graph by adding user attributes}
\usage{
EnhanceNetwork.twitter(dfCollect, dfRelations, dfUsers,
  lookupMissingUsers = TRUE, twitterAuth = NULL, writeToFile = FALSE)
}
\arguments{
\item{dfCollect}{A dataframe containing the collected tweet data from \code{Collect}.}

\item{dfRelations}{A dataframe containing the network relations data from \code{Create}.}

\item{dfUsers}{A dataframe containing the network users data from \code{Create}.}

\item{lookupMissingUsers}{Logical. Request user information for any users missing from dfCollect. Default 
is \code{TRUE}.}

\item{twitterAuth}{A twitter authentication object from \code{Authenticate}.}

\item{writeToFile}{Logical. If \code{TRUE} a data frame of user information and the resulting network graph will 
be saved to file. Default is \code{FALSE}.}
}
\value{
A list containing a dataframe with user information and an igraph object of the twitter network with 
user node attributes.
}
\description{
Creates a network from the relations and users dataframes generated by Create. Network is supplemented with 
additional downloaded user information applied as node attributes.
}
\note{
Only supports twitter actor network at this time. Bimodal network support will require the filtering 
of twitter user ids from nodes of other types.
}
