% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.youtube.R
\name{GetYoutubeVideoIDs}
\alias{GetYoutubeVideoIDs}
\title{Extract the ids from a list of youtube video urls}
\usage{
GetYoutubeVideoIDs(urls = NULL, file = NULL)
}
\arguments{
\item{urls}{Character vector. List of youtube urls.}

\item{file}{Character string. Text file containing youtube urls.}
}
\value{
A vector of youtube video ids as character strings that were extracted from input video urls.
}
\description{
This function reads youtube video urls from a list and or a text file and converts them to a vector of video ids. 
For example, url "https://www.youtube.com/watch?v=73I5dRucCds" returns the id "73I5dRucCds". This function can be 
used to create a vector for the youtube \code{Collect} methods \code{videoIDs} parameter.
}
\seealso{
\code{Collect}
}
\keyword{collect}
\keyword{id}
\keyword{video}
\keyword{youtube}
