% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportData}
\alias{ImportData}
\title{Import collected data from dataframe or previously saved to file}
\usage{
ImportData(data, socialmedia, type = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Character string or dataframe. Collected data file path or dataframe object.}

\item{socialmedia}{Character string. Social media type of collected data \code{twitter}, \code{youtube} or
\code{reddit}.}

\item{type}{Character string. Type of file or file format of file to import \code{csv} or \code{rds}. Default is
\code{NULL} to use extension.}

\item{verbose}{Logical. Output additional information. Default is \code{FALSE}.}
}
\value{
A dataframe with datasource and socialmedia class attributes.
}
\description{
Imports collected data from file or dataframe. Ensures \code{datasource} and specified
\code{socialmedia} type are set so data is usable by \code{\link{Create}} functions. Not required if collected data
was collected by \code{vosonSML} and saved as an \code{rds} file, use \code{\link{readRDS}} instead.
}
\examples{
\dontrun{
# import collected data from file
twitter_data <- ImportData("./data.csv", "twitter")

# import collected data from dataframe
twitter_data <- ImportData(rtweet_data, "twitter")
}

}
