% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulateUserInfoTwitter.R
\name{PopulateUserInfoTwitter}
\alias{PopulateUserInfoTwitter}
\title{Populate Twitter networks with user information}
\usage{
PopulateUserInfoTwitter(networkObject)
}
\arguments{
\item{networkObject}{An igraph graph object created with \code{\link{Create}}.}
}
\value{
An igraph graph object with additional user information.
}
\description{
This function is used to 'populate' Twitter networks (generated with the \code{\link{Create}} function) with 
information about the users in the network. This involves calls to the Twitter API to collect this information, 
which is then applied to the network as vertex attributes.
}
\examples{
\dontrun{
require(magrittr)

# get twitter user information and apply to network
myTwitNetwork_UserInfo <- PopulateUserInfoTwitter(myTwitNetwork)
}

}
\seealso{
\code{\link{Collect}}, \code{\link{Create}}
}
