% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SetOutputCat}
\alias{SetOutputCat}
\title{Set function output to use the \code{\link{cat}} method instead of \code{\link{message}}}
\usage{
SetOutputCat(flag = TRUE)
}
\arguments{
\item{flag}{Logical. Set the \code{voson.msg = "cat"} option. Set to \code{FALSE} or \code{NULL} to clear. Default is
\code{TRUE}.}
}
\description{
Set function output to use the \code{\link{cat}} method instead of \code{\link{message}}
}
