% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson-package.R
\docType{package}
\name{voson-package}
\alias{voson-package}
\alias{voson}
\title{voson: Collecting online data and generating networks for analysis}
\description{
A suite of easy to use functions for collecting online data and generating networks for analysis. Supports Mastodon, YouTube, Reddit and web hyperlink data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vosonlab/vosonSML}
  \item Report bugs at \url{https://github.com/vosonlab/vosonSML/issues}
}

}
\author{
\strong{Maintainer}: Bryan Gertzel \email{bryan.gertzel@anu.edu.au}

Authors:
\itemize{
  \item Robert Ackland \email{robert.ackland@anu.edu.au} (\href{https://orcid.org/0000-0002-0008-1766}{ORCID})
  \item Timothy Graham (\href{https://orcid.org/0000-0002-4053-9313}{ORCID})
}

Other contributors:
\itemize{
  \item Francisca Borquez [contributor]
}

}
\keyword{internal}
