% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddText.R
\name{vosonSML::AddText.actor.youtube}
\alias{vosonSML::AddText.actor.youtube}
\alias{AddText.actor.youtube}
\title{Add columns containing text data to youtube actor network dataframes}
\usage{
\method{AddText}{actor.youtube}(net, data, replies_from_text = TRUE,
  at_replies_only = TRUE, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{data}{A dataframe generated by \code{Collect}.}

\item{replies_from_text}{Logical. If comment text for an edge begins with \code{screen_name} change the
edge to be directed to \code{screen_name} - if different from the top level comment author that the reply
comment was posted to. Default is \code{TRUE}.}

\item{at_replies_only}{Logical. Comment \code{screen_names} must begin with an '@' symbol to be redirected.
Default is \code{TRUE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}
including columns containing text data.
}
\description{
Add columns containing text data to youtube actor network dataframes
}
