% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddText.R
\name{vosonSML::AddText}
\alias{vosonSML::AddText}
\alias{AddText}
\title{Add columns containing text data to network dataframes}
\usage{
AddText(net, data, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{data}{A dataframe generated by \code{Collect}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}
including columns containing text data.
}
\description{
Add columns containing text data to network dataframes
}
