% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddUserData.R
\name{vosonSML::AddUserData}
\alias{vosonSML::AddUserData}
\alias{AddUserData}
\title{Add columns of user information as node attributes to network dataframes}
\usage{
AddUserData(net, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{...}{Additional parameters passed to function.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}
including columns for additional user data.
}
\description{
Network is supplemented with additional downloaded social media user information applied as node
attributes.
}
\note{
Only supports twitter actor networks at this time. Refer to \code{\link{AddUserData.actor.twitter}}.
}
