% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVideoData.R
\name{vosonSML::AddVideoData.actor.youtube}
\alias{vosonSML::AddVideoData.actor.youtube}
\alias{AddVideoData.actor.youtube}
\title{Add columns of video information to youtube actor network dataframes}
\usage{
\method{AddVideoData}{actor.youtube}(net, youtubeAuth = NULL,
  videoIds = NULL, actorSubOnly = FALSE, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{youtubeAuth}{Youtube Authenticate object.}

\item{videoIds}{List. Video ID's to lookup.}

\item{actorSubOnly}{Logical. Only substitute video ID's for their publishers channel ID.
Don't add additional video data to edge list.}

\item{...}{Additional parameters passed to function.}
}
\value{
Network as a named list of three dataframes containing \code{$nodes}, \code{$edges}
and \code{$videos} nodes and edges include columns for additional video data.
}
\description{
Youtube actor network is supplemented with additional downloaded video information.
}
\examples{
\dontrun{
# add video data to an actor network
actorNetwork <- collectData \%>\% Create("actor") \%>\% AddVideoData(youtubeAuth)

# network
# actorNetwork$nodes
# actorNetwork$edges
# actorNetwork$videos
}

}
