% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVideoData.R
\name{vosonSML::AddVideoData}
\alias{vosonSML::AddVideoData}
\alias{AddVideoData}
\title{Add columns of video information to network dataframes}
\usage{
AddVideoData(net, youtubeAuth = NULL, videoIds = NULL, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{youtubeAuth}{Youtube Authenticate object.}

\item{videoIds}{List. Video ID's to lookup.}

\item{...}{Additional parameters passed to function.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}
including columns for additional video data.
}
\description{
Network is supplemented with additional downloaded video information.
}
