% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{vosonSML::graph}
\alias{vosonSML::graph}
\alias{graph}
\title{Create a graph from network}
\usage{
graph(network, directed = TRUE, write_file = FALSE, ...)
}
\arguments{
\item{network}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{directed}{Logical. Create a directed graph. Default is \code{TRUE}.}

\item{write_file}{Logical. Save graph to a file in the current working directory. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
An igraph object.
}
\description{
Create a graph from network
}
