% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.R
\name{Authenticate}
\alias{Authenticate}
\title{Create a credential to access social media APIs}
\usage{
Authenticate(socialmedia, ...)
}
\arguments{
\item{socialmedia}{Character string. Identifier for social media API to authenticate.\cr
Supports: \code{"twitter"}, \code{"youtube"}, \code{"reddit"}.}

\item{...}{Additional parameters for authentication appropriate to \code{socialmedia} identifier.
\describe{
  \item{twitter:}{\code{[appName], apiKey, apiSecret, accessToken, accessTokenSecret,}\cr
                  \code{[useCachedToken]}}
  \item{youtube:}{\code{apiKey}}
  \item{reddit:}{\code{No credentials required.}}
}}
}
\value{
A \code{credential} object with authentication information.
}
\description{
\code{Authenticate} creates a \code{credential} object that enables R to make authenticated calls to social media
APIs. A \code{credential} object is a S3 object with the authentication-related information such as access tokens
and the information on the social media that grant authentication. \code{Authenticate} is the first step of the
\code{Authenticate}, \code{\link{Collect}} and \code{\link{Create}} workflow.
}
\seealso{
\code{\link{Collect}}, \code{\link{Create}}
}
\keyword{authenticate}
\keyword{credential}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
