% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateWithTwitterAPI.R
\name{AuthenticateWithTwitterAPI}
\alias{AuthenticateWithTwitterAPI}
\title{Note: this function is DEPRECATED and will be removed in a future release.
Please use the \code{Authenticate} function}
\usage{
AuthenticateWithTwitterAPI(api_key, api_secret, access_token,
  access_token_secret, createToken)
}
\arguments{
\item{api_key}{character string specifying the 'API key' used for
authentication.}

\item{api_secret}{character string specifying the 'API secret' used for
authentication.}

\item{access_token}{character string specifying the 'access token' used for
authentication.}

\item{access_token_secret}{character string specifying the 'access token
secret' used for authentication.}

\item{createToken}{logical. !! NOT PROPERLY IMPLEMENTED YET.}
}
\value{
This is called for its side effect.
}
\description{
Twitter API Authentication
}
\details{
Oauth based authentication with the Twitter API

In order to collect data from Twitter, the user must first authenticate with
Twitter's Application Programming Interface (API).

This requires setting up an App on Twitter. An excellent guide to achieving
this can be found at:
http://thinktostart.com/twitter-authentification-with-r/
}
\examples{

\dontrun{
  # Firstly specify your API credentials
  my_api_key <- "1234567890qwerty"
  my_api_secret <- "1234567890qwerty"
  my_access_token <- "1234567890qwerty"
  my_access_token_secret <- "1234567890qwerty"

  AuthenticateWithTwitterAPI(api_key=my_api_key, api_secret=my_api_secret,
    access_token=my_access_token, access_token_secret=my_access_token_secret)
}
}
\seealso{
\code{AuthenticateWithFacebookAPI} and
\code{AuthenticateWithYouTubeAPI} for other ways to collect social media
data.
}
\author{
Timothy Graham <timothy.graham@anu.edu.au> & Robert Ackland
<robert.ackland@anu.edu.au>
}
\keyword{SNA}
\keyword{media}
\keyword{social}
\keyword{twitter}
