% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.R
\name{Collect}
\alias{Collect}
\title{Collect data from social media for generating networks}
\usage{
Collect(credential, ...)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate}.}

\item{...}{Additional parameters for data collection by appropriate to credential \code{socialmedia} type.
\describe{
  \item{twitter:}{\code{authToken, searchTerm, [searchType, numTweets, includeRetweets, retryOnRateLimit,}\cr
                  \code{writeToFile, verbose, ...]}}
  \item{youtube:}{\code{videoIDs, apiKeyYoutube, [verbose, writeToFile, maxComments]}}
  \item{reddit:}{\code{threadUrls, [waitTime, writeToFile]}}
}}
}
\value{
A data.frame object of class \code{dataSource.*} that can be used with \code{Create}.
}
\description{
This function collects data from social media APIs, and structures the data into a data frame of class
\code{dataSource.*}, ready for creating networks for further analysis. \code{Collect} is the second step of the
\code{Authenticate}, \code{Collect}, \code{Create} workflow. This function is a convenient UI wrapper to the core
CollectDataFrom* family of functions.
}
\seealso{
\code{Authenticate}, \code{Create}
}
\keyword{collect}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
