% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.web.R
\name{Collect.web}
\alias{Collect.web}
\title{Collect hyperlinks from web pages}
\usage{
\method{Collect}{web}(credential, pages, writeToFile = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{credential}{A \code{credential} object generated from \code{Authenticate} with class name \code{"web"}.}

\item{pages}{Dataframe. Dataframe of web pages to crawl.}

\item{writeToFile}{Logical. Write collected data to file. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the data collection. Default is \code{TRUE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
A \code{data.frame} object with class names \code{"datasource"} and \code{"web"}.
}
\description{
Collects hyperlinks from web pages and structures the data into a dataframe with the
class names \code{"datasource"} and \code{"web"}.
}
\examples{
\dontrun{
webData <- webAuth \%>\%
  Collect(pages = list(), writeToFile = TRUE)
}

}
