% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataReddit.R
\name{CollectDataReddit}
\alias{CollectDataReddit}
\title{Collect reddit thread data}
\usage{
CollectDataReddit(threadUrls, waitTime = 5, writeToFile)
}
\arguments{
\item{threadUrls}{character string vector. Reddit thread url's to collect data from.}

\item{waitTime}{numeric integer. Time in seconds to wait in-between url collection requests.}

\item{writeToFile}{logical. If the data should be written to file.}
}
\value{
A data frame object of class dataSource.reddit that can be used for creating unimodal 
networks (CreateActorNetwork).
}
\description{
Uses RedditExtractoR::reddit_content to collect user and comment data for thread urls.
}
\note{
The reddit API endpoint used for thread collection has maximum limit of 500 comments per thread url.
}
