% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectDataReddit.R
\name{collectDataReddit}
\alias{collectDataReddit}
\title{Collect reddit data}
\usage{
collectDataReddit(thread_urls, wait_time = 5, write_to_file)
}
\arguments{
\item{thread_urls}{Character string vector. Reddit thread url's to collect data from.}

\item{wait_time}{Numeric integer. Time in seconds to wait in-between url collection requests.}

\item{write_to_file}{Boolean. If the data should be written to file.}
}
\value{
A data frame object of class dataSource.reddit that can be used for creating unimodal 
networks (CreateActorNetwork).
}
\description{
Uses RedditExtractoR::reddit_content to collect user and comment data for thread urls.
}
\note{
The reddit API endpoint used for thread collection has maximum limit of 500 comments per thread url.
}
