% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.activity.twitter.R
\name{Create.activity.twitter}
\alias{Create.activity.twitter}
\title{Create twitter activity network}
\usage{
\method{Create}{activity.twitter}(datasource, type, writeToFile = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"twitter"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"activity"}.}

\item{writeToFile}{Logical. Save network data to a file in the current working directory. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Graph
}
\description{
Creates a twitter activity network from tweets.
}
\examples{
\dontrun{
# create a twitter activity network graph
}

}
