% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.R
\name{Create}
\alias{Create}
\title{Create networks from social media data}
\usage{
Create(dataSource, type = "actor", ...)
}
\arguments{
\item{dataSource}{A data frame of class \code{dataSource}.}

\item{type}{Character string. Type of network to be created, can be \code{"actor"}, \code{"bimodal"},
\code{"dynamic"}, \code{"semantic"} or \code{"ego"}.}

\item{...}{Additional parameters for network creation for appropriate \code{socialmedia} and network \code{type}.
Refer to Create*Network functions for more details.
\describe{
  \item{twitter|actor:}{\code{[writeToFile]}}
  \item{youtube|actor:}{\code{[writeToFile]}}
  \item{reddit|actor:}{\code{[weightEdges, includeTextData, cleanText, writeToFile]}}
  \item{twitter|bimodal:}{\code{[writeToFile, removeTermsOrHashtags]}}
  \item{instagram|bimodal:}{\code{[writeToFile, ...]}}
  \item{facebook|bimodal:}{\code{[writeToFile, removeTermsOrHashtags, ...]}}
  \item{facebook|dynamic:}{\code{[writeToFile]}}
  \item{twitter|semantic:}{\code{[writeToFile, termFreq, hashtagFreq, removeTermsOrHashtags,}\cr
                           \code{stopwordsEnglish]}}
  \item{instagram|ego:}{[writeToFile]}
}}
}
\value{
An \code{igraph} graph object.
}
\description{
This function creates networks from social media data (i.e. from data frames of class \code{dataSource}.
\code{Create} is the final step of the \code{Authenticate}, \code{Collect}, \code{Create} workflow. This function is
a convenient UI wrapper to the core Create*Network family of functions.
}
\details{
Note: when creating twitter networks, a network with additional user information can be generated using the
\code{\link{GraphUserInfoTwitter}} function. Additional calls can be made to the Twitter API to get information
about users that became nodes during network creation.
}
\examples{
\dontrun{
require(magrittr)

## youtube actor network example

myYoutubeAPIKey <- "xxxxxxxxxxxxxxxxxxxxxx"
listYoutubeVideoIDs <- c("W2GZFeYGU3s", "mL27TAJGlWc")

myActorNetwork <- Authenticate("youtube", apiKey = myYoutubeAPIKey) \%>\%
  Collect(videoIDs = listYoutubeVideoIDs) \%>\% Create("actor")

## instagram ego network example

myInstaAppID <- "xxxxxxxxxxx"
myInstaAppSecret <- "xxxxxxxxxxxxxxxxxxxxxx"
listInstaUsernames <- c("senjohnmccain", "obama")

myEgoNetwork <- Authenticate("instagram", appID = myInstaAppID, appSecret = myInstaAppSecret) \%>\%
  Collect(ego = TRUE, username = listInstaUsernames) \%>\% Create("ego")
}

}
\seealso{
\code{\link{Authenticate}}, \code{\link{Collect}}
}
\keyword{actor}
\keyword{bimodal}
\keyword{create}
\keyword{dynamic}
\keyword{ego}
\keyword{facebook}
\keyword{instagram}
\keyword{semantic}
\keyword{twitter}
\keyword{youtube}
