% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.R
\name{Create}
\alias{Create}
\title{Create networks from social media data}
\usage{
Create(datasource, type, ..., writeToFile = FALSE, verbose = TRUE)
}
\arguments{
\item{datasource}{Collected social media data of class \code{"datasource"} and \code{socialmedia}.}

\item{type}{Character string. Type of network to be created, can be \code{"activity"}, \code{"actor"},
\code{"twomode"} or \code{"semantic"}.}

\item{...}{Optional parameters to pass to functions providied by supporting R packages that are used for social media
network creation.}

\item{writeToFile}{Logical. Write data to file. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information. Default is \code{TRUE}.}
}
\description{
This function creates networks from social media data as produced from \code{\link{Collect}}.
\code{Create} is the final step of the \code{\link{Authenticate}}, \code{\link{Collect}} and \code{Create}
workflow.
}
