% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.R
\name{Create}
\alias{Create}
\title{Create networks from social media data}
\usage{
Create(dataSource, type = "actor", ...)
}
\arguments{
\item{dataSource}{Social media data collected using the \code{Collect} method.}

\item{type}{Character string. Type of network to be created, can be \code{actor}, \code{bimodal} or \code{semantic}.}

\item{...}{Additional parameters for network creation for appropriate \code{social media} and network \code{type}. 
Refer to S3 methods \code{social media} type for default parameters.}
}
\value{
Network data containing an igraph object.
}
\description{
This function creates networks from social media data (i.e. collected from dataframes of class \code{social media}).
\code{Create} is the final step of the \code{Authenticate}, \code{Collect}, \code{Create} workflow. This function 
is a wrapper for the Create*Network S3 methods.
}
\note{
When creating twitter networks, a network with additional user information can be generated using the
\code{\link{GraphUserInfoTwitter}} function. Additional calls can be made to the twitter API to get information
about users that were identified as nodes during network creation.
}
\seealso{
\code{\link{CreateActorNetwork}}, \code{\link{CreateBimodalNetwork}}, \code{\link{CreateSemanticNetwork}}
}
\keyword{actor}
\keyword{bimodal}
\keyword{create}
\keyword{network}
\keyword{semantic}
