% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.semantic.twitter.R
\name{Create.semantic.twitter}
\alias{Create.semantic.twitter}
\title{Create twitter semantic network}
\usage{
\method{Create}{semantic.twitter}(datasource, type,
  removeTermsOrHashtags = NULL, stopwordsEnglish = TRUE,
  termFreq = 5, hashtagFreq = 50, writeToFile = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}
}
\value{
A twitter semantic network as igraph object.
}
\description{
Creates a semantic network from collected tweets.
}
