% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.twomode.twitter.R
\name{Create.twomode.twitter}
\alias{Create.twomode.twitter}
\title{Create twitter twomode network}
\usage{
\method{Create}{twomode.twitter}(datasource, type, removeTermsOrHashtags = NULL, verbose = FALSE, ...)
}
\arguments{
\item{datasource}{Collected social media data with \code{"datasource"} and \code{"twitter"} class names.}

\item{type}{Character string. Type of network to be created, set to \code{"twomode"}.}

\item{removeTermsOrHashtags}{Character vector. Terms or hashtags to remove from the twomode network. For example, 
this parameter could be used to remove the search term or hashtag that was used to collect the data by removing any
nodes with matching name. Default is \code{NULL} to remove none.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to function. Not used in this method.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}.
}
\description{
Creates a twomode network from tweets returned from the twitter search query. In this network there are 
two types of nodes, twitter users who have tweeted (actors) and hashtags found within their tweets. Network edges 
are weighted and represent hashtag usage by the actor or specifically their tweets that contain a hashtag matching 
the name of the node they are directed towards.
}
\examples{
\dontrun{
# create a twitter twomode network graph removing the hashtag '#auspol' as it was used in 
# the twitter search query
twomodeNetwork <- twitterData \%>\% 
                  Create("twomode", removeTermsOrHashtags = c("#auspol"), verbose = TRUE)

# network
# twomodeNetwork$nodes
# twomodeNetwork$edges
}

}
