% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateActorNetwork.R,
%   R/CreateActorNetwork.reddit.R, R/CreateActorNetwork.twitter.R,
%   R/CreateActorNetwork.youtube.R
\name{CreateActorNetwork}
\alias{CreateActorNetwork}
\alias{CreateActorNetwork.reddit}
\alias{CreateActorNetwork.twitter}
\alias{CreateActorNetwork.youtube}
\title{Create an actor network from social media data}
\usage{
CreateActorNetwork(x, ...)

\method{CreateActorNetwork}{reddit}(x, weightEdges = FALSE,
  textData = FALSE, cleanText = TRUE, writeToFile = FALSE, ...)

\method{CreateActorNetwork}{twitter}(x, writeToFile = FALSE,
  verbose = FALSE, ...)

\method{CreateActorNetwork}{youtube}(x, writeToFile = FALSE, ...)
}
\arguments{
\item{x}{Collected social media data with \code{social media} class attribute.}

\item{...}{Additional parameters to pass to the network creation method.}

\item{weightEdges}{Logical. Combines and weights directed network edges. Default is \code{FALSE}.}

\item{textData}{Logical. If the igraph network should include the comment text as an edge attribute. 
Cannot be used with the \code{weightEdges} parameter. Default is \code{FALSE}.}

\item{cleanText}{Logical. If non-alphanumeric, non-punctuation, and non-space characters should be removed from the 
included text attribute data. Only applies if \code{textData = TRUE}. Default is \code{TRUE}.}

\item{writeToFile}{Logical. Save network data to a file in the current working directory. Default is \code{FALSE}.}

\item{verbose}{Logical. Output additional information about the network creation. Default is \code{FALSE}.}
}
\value{
A reddit actor network as igraph object.

A twitter actor network as list containing a relations dataframe, users dataframe and igraph object.

A youtube actor network as igraph object.
}
\description{
This function creates an actor network from social media data collected using the \code{Collect} method. Edges in 
the network represent interactions or relationships between the actors. For example, with twitter data an 
interaction is defined as a 'mention', reply' or 'retweet' from user i to user j, given 'tweet' m. With youtube 
comments, an interaction is defined as a 'reply' from user i to user j, given 'comment' m. The resulting network is 
returned as an igraph object.
}
\note{
Supported data sources: \code{twitter}, \code{youtube}, \code{reddit}
}
\seealso{
\code{\link{Create}}
}
\keyword{actor}
\keyword{create}
\keyword{network}
\keyword{reddit}
\keyword{twitter}
\keyword{youtube}
