% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateActorNetwork.reddit.R
\name{CreateActorNetwork.reddit}
\alias{CreateActorNetwork.reddit}
\title{Creates a reddit actor network from threads.}
\usage{
\method{CreateActorNetwork}{reddit}(x, writeToFile)
}
\arguments{
\item{x}{A dataframe as vosonSML class object containing collected social network data.}

\item{write_to_file}{Boolean. If the igraph network graph should be written to file.}
}
\value{
An igraph object of the actor network.
}
\description{
Uses RedditExtractoR::user_network to create an igraph actor network with posts as edge attribute.
}
