% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_youtube.R
\name{GetYoutubeVideoIDs}
\alias{GetYoutubeVideoIDs}
\title{Extract the ids from a list of YouTube video urls}
\usage{
GetYoutubeVideoIDs(urls = c(), verbose = FALSE)
}
\arguments{
\item{urls}{Character vector. List of YouTube urls or ids.}

\item{verbose}{Logical. Output additional information. Default is \code{FALSE}.}
}
\value{
A vector of YouTube video ids as character strings that were extracted from input video urls.
}
\description{
This function reads YouTube video urls from a list and or a text file and converts them to a vector of
video ids. For example, URL \code{https://www.youtube.com/watch?v=73I5dRucCds} returns the id \code{73I5dRucCds}.
This function can be used to create a vector for the YouTube \code{\link{Collect.youtube}} functions
\code{videoIDs} parameter.
}
\note{
Accepts YouTube url formats \code{https://youtu.be/xxxxxxxx} and
\code{https://www.youtube.com/watch?v=xxxxxxxx}.
}
