% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveCredential.R
\name{SaveCredential}
\alias{SaveCredential}
\alias{LoadCredential}
\title{Save and load credential information}
\usage{
SaveCredential(credential, filename)

LoadCredential(filename)
}
\arguments{
\item{credential}{A \code{credential} object.}

\item{filename}{Character string. Filename to be saved to or restored from. Default value is \code{credential.RDS}.}
}
\value{
A \code{credential} object.
}
\description{
Functions to save and load credential information. Currently, credential information will be stored as a RDS file. 
\code{SaveCredential} will return the input \code{credential}, useful for working as a filter between 
\code{Authenticate} and \code{Collect}.
}
\examples{
\dontrun{
require(magrittr)

## save credential example

myIgAppID <- "xxxxxxxxxxx"
myIgAppSecret <- "xxxxxxxxxxxxxxxxxxxxxx"
listIgUsernames <- c("senjohnmccain", "obama")

Authenticate("instagram", appID = myIgAppID, appSecret = myIgAppSecret) \%>\% 
  SaveCredential("instagramCred.RDS") \%>\% 
  Collect(ego = TRUE, username = listIgUsernames) \%>\% Create()

## load previously saved credential example

LoadCredential("instagramCred.RDS") \%>\% 
  Collect(tag = "obama", distance = 5000, n = 100) \%>\% Create("bimodal")
}

}
