% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetOutputCat}
\alias{SetOutputCat}
\title{Set function output to use the cat method instead of message}
\usage{
SetOutputCat(flag = TRUE)
}
\arguments{
\item{flag}{Logical. Set the \code{voson.msg = "cat"} option. Set to \code{FALSE} or \code{NULL} to
clear. Default is \code{TRUE}.}
}
\description{
Set function output to use the cat method instead of message
}
