% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVideoData.R
\name{vosonSML::AddVideoData.actor.youtube}
\alias{vosonSML::AddVideoData.actor.youtube}
\alias{AddVideoData.actor.youtube}
\title{Add columns of video information to youtube actor network dataframes}
\usage{
\method{AddVideoData}{actor.youtube}(net, youtubeAuth = NULL,
  videoIds = NULL, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{youtubeAuth}{Youtube Authenticate object.}

\item{videoIds}{List. Video ID's to lookup.}

\item{...}{Additional parameters passed to function.}
}
\value{
Network as a named list of two dataframes containing \code{$nodes} and \code{$edges}
including columns for additional video data.
}
\description{
Youtube actor network is supplemented with additional downloaded video information.
}
