% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVideoData.R
\name{vosonSML::AddVideoData}
\alias{vosonSML::AddVideoData}
\alias{AddVideoData}
\title{Add columns of video information to network dataframes}
\usage{
AddVideoData(net, youtubeAuth = NULL, ...)
}
\arguments{
\item{net}{A named list of dataframes \code{nodes} and \code{edges} generated by \code{Create}.}

\item{youtubeAuth}{Youtube Authenticate object.}

\item{...}{Additional parameters passed to function.}
}
\value{
Network as a named list of three dataframes containing \code{$nodes}, \code{$edges}
and \code{$videos} nodes and edges include columns for additional video data.
}
\description{
Network is supplemented with additional downloaded video information.
}
\note{
Only supports youtube actor networks. Refer to \code{\link{AddVideoData.actor.youtube}}.
}
