% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.youtube.R
\name{vosonSML::GetYoutubeVideoIDs}
\alias{vosonSML::GetYoutubeVideoIDs}
\alias{GetYoutubeVideoIDs}
\title{Extract the ids from a list of youtube video URLs}
\usage{
GetYoutubeVideoIDs(urls = NULL, file = NULL)
}
\arguments{
\item{urls}{Character vector. List of youtube URLs.}

\item{file}{Character string. Text file containing youtube URLs.}
}
\value{
A vector of youtube video ids as character strings that were extracted from input video urls.
}
\description{
This function reads youtube video urls from a list and or a text file and converts them to a vector of 
video ids. For example, URL \code{https://www.youtube.com/watch?v=73I5dRucCds} returns the id \code{73I5dRucCds}. 
This function can be used to create a vector for the youtube \code{\link{Collect.youtube}} functions \code{videoIDs} 
parameter.
}
\note{
Accepts youtube URL formats \code{https://youtu.be/xxxxxxxx} and 
\code{https://www.youtube.com/watch?v=xxxxxxxx}.
}
