% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{vosonSML::ImportData}
\alias{vosonSML::ImportData}
\alias{ImportData}
\title{Import collected data previously saved to file}
\usage{
ImportData(file, type = "csv", datasource)
}
\arguments{
\item{file}{Character string. Collected data file path.}

\item{type}{Character string. Type of file or file format of file to import.}

\item{datasource}{Character string. Social media type of collected data.}
}
\value{
A dataframe with datasource class attributes usable by create network functions.
}
\description{
Import collected data previously saved to file
}
\seealso{
\code{\link{Collect}}
}
