% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML-package.R
\docType{package}
\name{vosonSML-package}
\alias{vosonSML-package}
\alias{_PACKAGE}
\alias{vosonSML}
\title{vosonSML: Collecting Social Media Data and Generating Networks for Analysis}
\description{
A suite of easy to use functions for collecting social media data and generating networks for analysis. Supports Mastodon, YouTube, Reddit, Twitter and web site data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vosonlab/vosonSML}
  \item Report bugs at \url{https://github.com/vosonlab/vosonSML/issues}
}

}
\author{
\strong{Maintainer}: Bryan Gertzel \email{bryan.gertzel@anu.edu.au}

Authors:
\itemize{
  \item Robert Ackland \email{robert.ackland@anu.edu.au} (\href{https://orcid.org/0000-0002-0008-1766}{ORCID})
  \item Timothy Graham (\href{https://orcid.org/0000-0002-4053-9313}{ORCID})
}

Other contributors:
\itemize{
  \item Francisca Borquez [contributor]
}

}
\keyword{internal}
